requirements.txt
asgiref==3.7.2
asttokens==2.4.1
beautifulsoup4==4.12.2
crispy-bootstrap4==2023.1
crispy-forms-bootstrap2==2022.1
decorator==5.1.1
Django==4.2.7
django-bootstrap==0.2.4
django-bootstrap4==23.3
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
exceptiongroup==1.1.3
executing==2.0.1
flake8==6.1.0
flake8-builtins==2.2.0
flake8-print==5.0.0
install==1.3.5
ipython
jedi==0.19.1
matplotlib-inline==0.1.6
mccabe==0.7.0
parso==0.8.3
pexpect==4.8.0
prompt-toolkit==3.0.40
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.16.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
typing_extensions==4.8.0
wcwidth==0.2.9
Celery
requests
django-filter
djangorestframework
djangorestframework-simplejwt
pytest
pytest-django
pytest-cov
pytest-mock